3/15/22

+++ src/main/java/com/springboot/libraryOrg/repository/BooksRepository.java
@@ -8,7 +8,7 @@ import com.springboot.libraryOrg.models.Books;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called booksRepository
 
 @Repository
-public interface BooksRepository extends JpaRepository<Books, Integer> {
+public interface BooksRepository extends JpaRepository<Books, Long> {
 	
 }
